+++ kvirt/config.py
@@ -20,7 +20,7 @@ from time import sleep
 import webbrowser
 import yaml
 
-__version__ = '7.15'
+__version__ = '7.16'
 
 
 class Kconfig: